+++ backend/.gitignore
@@ -1,3 +1,3 @@
-.idea/*
+.idea/
 bin/
 obj/
