+++ backend/Nebulift.sln
@@ -1,6 +1,7 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nebulift", "Nebulift.csproj", "{563C6650-745E-4AEA-814A-2684932E1A71}"
+﻿Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nebulift.tests", "tests\Nebulift.tests\Nebulift.tests.csproj", "{3CD9C6D1-6907-407D-A368-D224BC823C00}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nebulift.Api", "src\Nebulift.Api\Nebulift.Api.csproj", "{D22629A0-2B5C-4A95-A7F1-AD4C2541C53E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -8,9 +9,13 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{563C6650-745E-4AEA-814A-2684932E1A71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{563C6650-745E-4AEA-814A-2684932E1A71}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{563C6650-745E-4AEA-814A-2684932E1A71}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{563C6650-745E-4AEA-814A-2684932E1A71}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3CD9C6D1-6907-407D-A368-D224BC823C00}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3CD9C6D1-6907-407D-A368-D224BC823C00}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3CD9C6D1-6907-407D-A368-D224BC823C00}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3CD9C6D1-6907-407D-A368-D224BC823C00}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D22629A0-2B5C-4A95-A7F1-AD4C2541C53E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D22629A0-2B5C-4A95-A7F1-AD4C2541C53E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D22629A0-2B5C-4A95-A7F1-AD4C2541C53E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D22629A0-2B5C-4A95-A7F1-AD4C2541C53E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
