+++ backend/src/Nebulift.Api/Nebulift.Api.http
@@ -1,0 +1,6 @@
+@Nebulift.Api_HostAddress = http://localhost:5052
+
+GET {{Nebulift.Api_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
