+++ backend/src/Nebulift.Api/Program.cs
@@ -34,8 +34,3 @@ public static class Program
         app.Run();
     }
 }
-
-public record WeatherForecast(DateOnly date, int temperatureC, string? summary)
-{
-    public int TemperatureF => 32 + (int)(temperatureC / 0.5556);
-}
