+++ backend/src/Nebulift.Api/appsettings.Development.json
